requirements.txt
pandas
numpy
pybids
mne
mne-bids
pyyaml
git+https://github.com/yjmantilla/sovabids.git#egg=sovabids
autoreject
mne-icalabel
pyprep
scipy
neuroCombat
mlxtend
xgboost
scikit-learn
matplotlib
plotly
tqdm
shap
psutil